xtask: require kernel flash/RAM to be 0 mod 4
Our linker scripts all assume that things are aligned to at least 32-bit
boundaries. Right now, if you configure an lpc55 target (at least) with
a kernel flash size that isn't a multiple of 4, you get an obscure error
from the linker late in the build process.

That's annoying because (1) it happens late, so iteration time is slow,
and (2) linker errors are designed by space aliens to confuse humanity,
or so I've inferred.

This adds an explicit check and early failure for the case.