sensor: validate `SensorId`s on the client side (#1683)
The `SensorId` type represents an index into an array of sensors in the
`sensors` server task. Currently, this is validated by the sensors
server, returning `SensorError::InvalidSensor` if it is out of range.
However, the sensor IDs used by firmware code are hard-coded by code
generation and should never be out of range, so in most cases, an
invalid `SensorId` should not occur and should be considered a bug. Only
`host-spc-comms` will ever request sensor IDs that are out of range, if
the upstack code requests an ID that doesn't exist on this device.

Therefore, we ought to change the sensor API to use reply-fault on
receipt of an invalid `SensorId`, rather than returning an error. This
commit prepares for such a change by adding client-side validation of
`SensorId`s with a `TryFrom` impl, and a panicking `SensorId::new` This
way, the tasks that *can* handle sensor IDs from the outside world
(`host-sp-comms`) can validate that sensor IDs received from the outside
world are valid _before_ calling into the sensor task API, and reject
any invalid IDs. In a subsequent PR, `task_sensor_api` will be updated
to remove the `InvalidSensor` error type.

The `host-sp-messages` crate is changed to replace the use of `SensorId`
with `u32`, as it deals with _un-validated_ sensor indices received from
the host over IPCC. `host-sp-comms` will validate these sensor IDs
before calling into sensor IPC methods.

A nice side benefit of using the panicking `const fn new` in generated
code is that...it's const fn, and the code generation mostly uses it in
`const` initializers. So, if any of the code generation accidentally
emits an invalid sensor ID, we'll actually get a build-time rather than
runtime error, which seems nice!