stm32xx-sys: configure GPIO IRQ edge with an enum
The current `Sys::gpio_irq_configure` IPC interface is a bit confusing,
as it takes two `bool`s for the rising and falling edge sensitivities,
as in:

```rust
sys.gpio_irq_configure(mask, true, false);
```

This isn't particularly clear for a reader, who then has to remember the
order of the arguments to determine which edge sensitivity is being
enabled here. This is a case of ["boolean blindness"][1].

This commit improves the interface by changing it to take an `enum` with
variants for `Rising`, `Falling`, and `Both`. Now, the same call to the
IPC looks like this:

```rust
sys.gpio_irq_configure(mask, Edge::Rising);
```

which should be much more obvious to a reader, who no longer needs to
remember the argument ordering to decypher calls to this function.

Also, this has the advantage of preventing the case where both rising
and falling edge sensitivities are disabled at compile-time. Currently,
this is a runtime error (reply fault), but with the new interface, a
caller can provide an invalid configuration if they hand-write the input
buffer rather than using the Idol client stub, so it should be harder to
mess up.

[1]: https://runtimeverification.com/blog/code-smell-boolean-blindness