thermal: remove panic formatting and `Debug` impls
Currently, the `thermal` task uses `unwrap()` in a bunch of places,
requiring it to derive `Debug` for a bunch of stuff. This commit changes
those to `unwrap_lite()`, and removes many of the derived `Debug` impls.
On `gimlet-f`, this shaves off about 1K of flash:

before:
```

thermal              flash   23316  23328  32768
                     ram     7820   8192   8192
```

after:

```
thermal              flash   22392  22400  32768
                     ram     7820   8192   8192
```

It would be nice to also remove the derived `Debug` from types like
`SensorId`, but it looks like that's currently used by `validate_api`
--- I wasn't sure if any of the `validate_api` types are used in
control-plane comms, so I didn't touch that yet. May do so in a
follow-up.