silence the clippy lint for "this entire thing" (#1740)
Clippy (quite rightly!) [complains] about code which "[c]reat[es] a
mutable reference which can be repeatedly derived from an immutable
reference", as this allows mutable aliasing. But, what Clippy *doesn't*
realize, because it's just a dumb machine that can only look at the
function's signature and apply simple rules to it, is that the entire
point of this thing is that the mutable reference can only be ever
created once. So, let's silence Clippy's criticism here.

[complains]:
    https://rust-lang.github.io/rust-clippy/master/index.html#/mut_from_ref