kerncore: Add tests for bad or uncontained memory contained in a contiguous slice (#1747)
The recent Oxide and Friends episode mentioned this memory region
spanning slices issue, and the code was interesting. I thought I saw a
chance to optimise it by omitting the start address containment checking
once the start region is found. (Based on benchmarking against the test
impls, it isn't faster.) The obvious issue with that was of course
validity of the intermediate regions but I realised that wouldn't be a
problem since I still need to check the validity of each region between
start and end.

I figured I'd write tests to show that the alternative solution works
but then realised the actual issue: My "optimised" algorithm allows for
uncontained (unmapped?) memory to be accessed as long as it is found
between two good regions. I presume this makes my improvement worthless,
but at least the tests are worthwhile additions. The second added test
shows the error in my idea.

I've still left the original modification in this PR's commits to at
least show what the idea was. There may still be some worth to it.

Side note: The idea for the two loops came from `Vec::retain_mut`'s code
which uses a similar logic to find the first item that isn't retained.

---------

Co-authored-by: Cliff L. Biffle <cliff@oxide.computer>