stm32h7-sprot-server: fix timer handling
Previously a race condition between the RoT IRQ event and the timer
event could cause this to detect a timeout immediately on the _next_
time it was called, due to a notification being left pending. This
is because the code assumed that the notification for the timer was
sufficient to give up on the remote device; that's not actually the
case.

This changes the code to look at the clock before making decisions about
elapsed time.