Enable notifications on any closed RECV.
Previously, a closed RECV against the kernel was special-cased as a
"listen only for notifications" operation. Closed RECV against any other
task would ignore the provided notification mask and treat it as zero.

This seemed weird and limiting.

This commit alters the semantics of RECV with a non-zero notification
mask and enables notifications in that specific situation.

As far as I can tell, we have no code that relies on notifications being
filtered out during closed-RECV, so this should be safe.

Fixes #1761.