Add support for {Read,Write}NoAddrIncr SPI operations, leverage them in `transceivers` (#1768)
Part 1) Adjustments to `fpga-api/src/lib.rs`, `fpga-server/src/main.rs`,
and `fpga.idol` for the new SPI operations.

As of https://github.com/oxidecomputer/quartz/pull/139 the FPGA's SPI
peripheral supports these operations. Prior to this PR the FPGA drivers
did not take an operation argument for read operations because there was
only a single type of read. Now that there are two (`Read` and
`ReadAddrNoIncr`), the intended operation must be specified. Given that
up until this point `Read` was the only option (and will probably remain
the primary use 99.9% of the time) that has been made the default for
the `read` operation which the majority of callers use. To utilize the
new read operation, one can just call the underlying `read_bytes`
function which takes a `ReadOp` parameter.

Part 2) Making use of the new SPI operations in the `transceivers` code

This also implements their use for transceivers I2C buffers for the QSFP
FPGA design. All the `*_regs.*` files are generated from the FPGA build
process and are brought in alongside the `.bit` file. The transceivers
code has been reworked to leverage the SPI refactoring from
https://github.com/oxidecomputer/quartz/pull/146. This code utilizes the
non-address incrementing SPI operations to read/write to the I2C buffers
in the FPGA.