jefe: expose armv8m cfg to fix bug
Jefe had a check that attempted to avoid building dump support on LPC55,
but that check didn't work, because it was checking for the presence of
one of our custom cfgs...but not doing the invocation in its build
script that actually generates that cfg.

rustc will start checking such things soon, which is how I found this.