Bump Idol to get foo_ARGS struct fix.
Between 46fec5217 (inclusive) and this commit (exclusive) we haven't
generated DWARF info for the Idol ARGS struct in the specific case of a
zero-argument function using non-zerocopy encoding. Humility 0.11.6 has
been fixed to tolerate this, but in the interest of broader
compatibility, we've also patched Idol to try and bring the debug info
back.

This pulls in the Idol fix.

Fixes #1777