psc-seq: don't repeat StickyFault every time.
We're currently in a situation where the dogfood rack has the power whip
feeding PSUs 3-5 disconnected. This means the PSUs are sitting there in
faulted state refusing to enable. Since we can't distinguish this from
actual PSU failures, currently we're frantically chatting about this
situation in the logs. This is annoying, and overwrites other data.

This change alters the behavior to note a sticky fault condition _once_
and then quietly keep trying to resolve it without the spam.