Adds a Medusa image (#1806)
The primary purpose of this PR is to sketch out an initial stab at an
image for Medusa, our QSFP front IO test fixture (details on that in RFD
405). The Medusa-specific logic lives in `medusa-seq-server`. The
current goal is just to support hardware bringup, and thus that
sequencer just attempts to power on the front IO board and then program
its FPGAs. It also checks the status of the various power rails on
Medusa and exposes an interface to control those rails (on Sidecar this
functionality is handled by the FPGA).

An adjacent piece of work here is that Medusa does not have any fans to
control and that broke the previous mold of "everything with `thermal`
has fans". `thermal` logs `sensor` data _and_ wants to control fan
speeds off that data. So the Medusa thermal BSP looks a bit odd compared
to the others. Additionally, the `transceivers` task had a
`thermal-control` feature flag added to gate the interactions with
`thermal` (Sidecar will have this feature set, Medusa will not).

Lastly, there's some ugliness that I did for the sake of expediency
given my personal situation. Until I can actually land #1805, Medusa
simply copies some code from Sidecar. Once this PR lands, I will extend
#1805 to cleanup Medusa as well.

Closes #1571

---------

Co-authored-by: Cliff L. Biffle <cliff@oxide.computer>