stm32xx-sys: move GPIO IRQ codegen to tasks
Using GPIO interrupts (EXTI) on STM32xx devices involves code generation
based on the pins mapped to EXTI interrupts in the app config.
Currently, this code generation is done when building the
`drv-stm32xx-sys-api` crate. Unfortunately, this codegen depends on
`task_name` to determine which task's GPIO pin IRQs to build the
generated code for, which results in `drv-stm32xx-sys-api` being
recompiled for each task in an image when it would not otherwise be
necessary to do so. This is a shame, as it results in substantially
longer builds; see #1828 for details.

There isn't *really* any reason that this code needs to be generated in
the `drv-stm32xx-sys-api` crate rather than in the `build.rs` files for
the tasks that actually use GPIO interrupts. The only reason for doing
it this way is "I thought it seemed nicer" and that it let downstream
tasks avoid typing
```rust
include!(concat!(env!("OUT_DIR"), "/gpio_irq_pins.rs"));
```
but that's not as important as avoiding unnecessary recompilation.
Therefore, this commit fixes this by removing the GPIO interrupt codegen
from `drv-stm32xx-sys-api`, and instead having the build scripts for the
individual tasks that actually require EXTI GPIO IRQs depend on
`build-stm32xx-sys` and do the codegen themselves. This way, we should
be able to avoid rebuilding `drv-stm32xx-sys-api` as many times.

This approach only required one very small change to the codegen,
changing it to import a few types from `drv-stm32xx-sys-api` rather than
`drv-stm32xx-gpio-common`, so that we don't have to add an explicit
dependency on the latter in the tasks that use the codegen. This
shouldn't have any real impact, since the only reason we got those types
from `drv-stm32xx-gpio-common` previously was that the generated code
was *in* `drv-stm32xx-sys-api`.

Fixes #1828