chore: Cleanup userlib namespace imports (#1855)
Didn't quite have the time to get rid of all of these, but I removed a
bunch of userlib namespace imports.

These were mentioned in a PR nearly half a year ago and I "helpfully"
went through and pointed out those that were in the PR, so this is my
atonement.

Each task is done in a separate commit here. All others except
`drv-lpc55-spi-server` is tested to cause no build errors, or I think
anyway. I'm randomly building tasks based on searches into the
repository with the device name, and testing that they fail if I remove
the `userlib::*` import entirely, and then add imports until build
works.