static-cell: add `StaticCell::try_borrow_mut` (#1856)
Presently, `StaticCell`s are accessed using the `StaticCell::borrow_mut`
method, which attempts to claim the cell, and panics if it is already
borrowed. This commit introduces a `try_borrow_mut` method, which
instead returns an `Option` that's `None` if the cell is already
borrowed. This is intended to allow `StaticCell` to be used in
non-panicking tasks.