MVP implementation of `HostFlash` for Cosmo (#1851)
This PR implements the `HostFlash` API for the Cosmo / Grapefruit flash
IC, mediated by an FPGA.

`gimlet-hf` is renamed to `hf` in a few places, because we're now using
the same API for both Gimlet and Cosmo systems. There are no changes to
the API itself.

`drv-fmc-nor-flash` is renamed to `drv-cosmo-hf` for symmetry with
Gimlet. It is split into `main.rs` (implementing the actual chip
wrangling) and `hf.rs` (implementing the `HostFlash` Idol API).

Many functions are left as stubs, to be implemented in later PRs.
However, this PR is sufficient to program and verify the flash chip with
`humility qspi`!

---------

Co-authored-by: Nathanael Huffman <nathanael@oxidecomputer.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>