Drop release
We were putting releases on GitHub but this doesn't quite work
with our current flow as we have extra steps to do after a build.
Drop this for now and we can bring it back later if necessary.

If we do decide to bring this back, we need to pickup the fix
in https://github.com/advisories/GHSA-cxww-7g56-2vh6