kern: handle pow2-sized stacks in small tasks
Right now, if a task has no non-stack RAM usage and a power-of-two
stack, the kernel's stack scribbling code skips it. This causes humility
stackmargin to emit bogus values, but has no other implications at the
moment.

The cause is a classic off-by-one error, testing the address just above
the stack instead of within the bounds of the stack. This hits a valid
region if there is at least one byte of non-stack data in RAM, but only
then.

This fixes that by testing the top word in the stack.

Fixes #1876.