Make `oxide-rot-1-dev` a completely different board
`oxide-rot-1/app.toml` expects to be run on a machine that has been
through the full manufacturing flow. If it is run on one that hasn't
had that happen it will fail to boot (loop forever). This is mostly
a concern for engineering development but it keeps tripping people up
so attempt to differentiate the correct image to use by changing the
board. It's not strictly incompatible but going from one to another kind
is a sort of cheap and relatively easy to rework.