Sidecar: Update FPGA to Quartz 8a2c75a (#1901)
Pulling in an update to the Sidecar mainboard FPGA which adds a debounce
to the PCIe PERST signal from the host. Details in
https://github.com/oxidecomputer/quartz/commit/8a2c75af0c51a7dd3e5ca2c34308dbbe72d4354e.

There's a significant amount of code deletion from the register map HTML
because I fixed a template bug that was generating a bunch of junk table
rows. Those are gone now!