Revert to hubtools 0.4.6, don't build with a default epoch value.
Don't require updating to hubtools 0.4.7. There doesn't need to be
an 'EPOC' in the caboose as it is always "0" at this step in
the release engineering flow.

Fix complementary swapped args and corresponding swapped comparison to
read correctly.

Remove vestigial epoch header check. With the decision to not use
ImageHeader.epoch, a TBD prep_image_update variant will contain
the epoch information.