kern: suppress phantom SVCs on ARMv6-M
You might recognize this as #1134. Well, it's back. My fix for the
problem on ARMv6-M was apparently never quite correct on M0+, because
the core register we were relying on to detect and cancel a pending SVC
-- SHCSR -- is _implemented_ but _only visible from the debugger._ So
when I tried it by hand in openocd, it worked... but from the kernel it
silently fails.

ARM, you are not making friends with stuff like this.

I've worked out a different way to _detect_ the "fault during SVC" case
using the always-implemented ICSR register. That register doesn't
provide any way to _cancel_ the SVC, so I've added conditional code to
the syscall entry path that can do it.

Fixes #1928.